refactor: standardize entry points
